deploy-docker/requirements.txt
click==8.0.4
Flask==2.2.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.1
Werkzeug==2.2.2
pandas==1.5.3
numpy==1.24.3
scikit-surprise==1.1.3